+++ IsekaiMod/Content/Features/IsekaiProtagonist/Archetypes/GodEmperor/BarrierSelection.cs
@@ -2,27 +2,62 @@
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Selection;
+using Kingmaker.Designers.EventConditionActionSystem.Actions;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
 using Kingmaker.ResourceLinks;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
+using Kingmaker.UnitLogic.Abilities.Components.AreaEffects;
 using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.UnitLogic.Buffs.Components;
 using Kingmaker.UnitLogic.FactLogic;
+using Kingmaker.UnitLogic.Mechanics.Actions;
 using Kingmaker.UnitLogic.Mechanics.Components;
+using Kingmaker.UnitLogic.Mechanics.Conditions;
 using Kingmaker.Utility;
 using TabletopTweaks.Core.Utilities;
+using UnityEngine;
 using static IsekaiMod.Main;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.GodEmperor {
 
     internal class BarrierSelection {
 
+        private static readonly Sprite Icon_GoldBarrier = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_AURA_BARRIER_GOLD.png");
+        private static readonly Sprite Icon_VoidBarrier = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_AURA_BARRIER_VOID.png");
+
+        private static readonly BlueprintBuff HeroismGreaterBuff = BlueprintTools.GetBlueprint<BlueprintBuff>("b8da3ec045ec04845a126948e1f4fc1a");
+
         public static void Add() {
+
+            // Extra effects for Hero progression
+            var GoldBarrierHeroism = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "GoldBarrierHeroism", bp => {
+                bp.SetName(IsekaiContext, "Greater Gold Barrier");
+                bp.SetDescription(IsekaiContext, "At 10th level, Gold Barrier now also grants Greater Heroism to your allies.");
+                bp.m_Icon = Icon_GoldBarrier;
+            });
+            var GoldBarrierFastHealing = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "GoldBarrierFastHealing", bp => {
+                bp.SetName(IsekaiContext, "Grand Gold Barrier");
+                bp.SetDescription(IsekaiContext, "At 12th level, Gold Barrier now also grants fast healing equal to your character level.");
+                bp.m_Icon = Icon_GoldBarrier;
+            });
+            var GoldBarrierFastHealingBuff = TTCoreExtensions.CreateBuff("GoldBarrierFastHealingBuff", bp => {
+                bp.SetName(IsekaiContext, "Grand Gold Barrier");
+                bp.SetDescription(IsekaiContext, "This character has fast healing equal to your character level.");
+                bp.m_Icon = Icon_GoldBarrier;
+                bp.AddComponent<AddEffectFastHealing>(c => {
+                    c.Heal = 0;
+                    c.Bonus = Values.CreateContextRankValue(AbilityRankType.StatBonus);
+                });
+                bp.AddComponent<ContextRankConfig>(c => {
+                    c.m_Type = AbilityRankType.StatBonus;
+                    c.m_BaseValueType = ContextRankBaseValueType.CharacterLevel;
+                });
+            });
+
             const string GoldBarrierName = "Gold Barrier";
             const string GoldBarrierDescription = "Allies within 40 feet of you gain a sacred bonus to AC and saving throws equal to 1/2 your character level.";
             const string GoldBarrierDescriptionBuff = "This character has a sacred bonus to AC and saving throws equal to 1/2 your character level.";
-            var Icon_GoldBarrier = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_AURA_BARRIER_GOLD.png");
             var GoldBarrierBuff = TTCoreExtensions.CreateBuff("GoldBarrierBuff", bp => {
                 bp.SetName(IsekaiContext, GoldBarrierName);
                 bp.SetDescription(IsekaiContext, GoldBarrierDescriptionBuff);
@@ -59,7 +94,61 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.GodEmperor {
                 bp.Shape = AreaEffectShape.Cylinder;
                 bp.Size = new Feet(40);
                 bp.Fx = new PrefabLink();
-                bp.AddComponent(AuraUtils.CreateUnconditionalAuraEffect(GoldBarrierBuff.ToReference<BlueprintBuffReference>()));
+                bp.AddComponent<AbilityAreaEffectRunAction>(c => {
+                    c.UnitEnter = Helpers.CreateActionList(
+                        new ContextActionApplyBuff {
+                            m_Buff = GoldBarrierBuff.ToReference<BlueprintBuffReference>(),
+                            Permanent = true,
+                            DurationValue = Values.Duration.Zero,
+                        },
+                        new Conditional {
+                            ConditionsChecker = ActionFlow.IfSingle<ContextConditionCasterHasFact>(c => {
+                                c.m_Fact = GoldBarrierFastHealing.ToReference<BlueprintUnitFactReference>();
+                                c.Not = false;
+                            }),
+                            IfTrue = ActionFlow.DoSingle<ContextActionApplyBuff>(c => {
+                                c.m_Buff = GoldBarrierFastHealingBuff.ToReference<BlueprintBuffReference>();
+                                c.Permanent = true;
+                                c.DurationValue = Values.Duration.Zero;
+                            }),
+                            IfFalse = ActionFlow.DoNothing()
+                        },
+                        new Conditional {
+                            ConditionsChecker = ActionFlow.IfSingle<ContextConditionCasterHasFact>(c => {
+                                c.m_Fact = GoldBarrierHeroism.ToReference<BlueprintUnitFactReference>();
+                                c.Not = false;
+                            }),
+                            IfTrue = ActionFlow.DoSingle<ContextActionApplyBuff>(c => {
+                                c.m_Buff = HeroismGreaterBuff.ToReference<BlueprintBuffReference>();
+                                c.Permanent = true;
+                                c.DurationValue = Values.Duration.Zero;
+                            }),
+                            IfFalse = ActionFlow.DoNothing()
+                        }
+                        );
+                    c.UnitExit = Helpers.CreateActionList(
+                        new ContextActionRemoveBuff {
+                            m_Buff = GoldBarrierBuff.ToReference<BlueprintBuffReference>(),
+                            RemoveRank = false,
+                            ToCaster = false,
+                            OnlyFromCaster = true
+                        },
+                        new ContextActionRemoveBuff {
+                            m_Buff = GoldBarrierFastHealingBuff.ToReference<BlueprintBuffReference>(),
+                            RemoveRank = false,
+                            ToCaster = false,
+                            OnlyFromCaster = true
+                        },
+                        new ContextActionRemoveBuff {
+                            m_Buff = HeroismGreaterBuff.ToReference<BlueprintBuffReference>(),
+                            RemoveRank = false,
+                            ToCaster = false,
+                            OnlyFromCaster = true
+                        }
+                        );
+                    c.UnitMove = ActionFlow.DoNothing();
+                    c.Round = ActionFlow.DoNothing();
+                });
             });
             var GoldBarrierAreaBuff = TTCoreExtensions.CreateBuff("GoldBarrierAreaBuff", bp => {
                 bp.SetName(IsekaiContext, GoldBarrierName);
@@ -90,7 +179,6 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.GodEmperor {
             const string VoidBarrierName = "Void Barrier";
             const string VoidBarrierDescription = "Allies within 40 feet of you gain a profane bonus to AC and saving throws equal to 1/2 your character level.";
             const string VoidBarrierDescriptionBuff = "This character has a profane bonus to AC and saving throws equal to 1/2 your character level.";
-            var Icon_VoidBarrier = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_AURA_BARRIER_VOID.png");
             var VoidBarrierBuff = TTCoreExtensions.CreateBuff("VoidBarrierBuff", bp => {
                 bp.SetName(IsekaiContext, VoidBarrierName);
                 bp.SetDescription(IsekaiContext, VoidBarrierDescriptionBuff);